+++ src/main/kotlin/model/InverterData.kt
@@ -3,7 +3,7 @@ package model
 data class InverterData (
     val inverterId: Int,
     val powerPlantId: Int,
-    val maxAllowedAcPower: Float,
+    val maxAllowedAcPower: Double,
     val controllable: Boolean,
     val type: InverterType
 )
